+++ CHANGELOG.md
@@ -1282,3 +1282,12 @@ All notable changes to this project will be documented in this file.
 **Logged At:** 2025-11-22T12:34:46+08:00
 
 ---
+
+### MA-NPM-START-MVP-MARKETING-UI-AUTOPILOT-BOOTSTRAP-2025-11-24T15:32:20+08:00
+**Agent Update (MVP - Brain, DB_Admin, Workflow, QA, UI, Designer)**
+
+- npm start: MVP Marketing UI + autopilot bootstrap
+
+**Logged At:** 2025-11-24T15:32:20+08:00
+
+---